refactor: Tmdb 컨텐츠 검색 리스트 응답을 ContentPageResponse로 변경하고 TmdbSearchDtoMapper 사용 방식으로 로직 변경

+++ src/main/java/net/watchbox/global/tmdb/inner/search/TmdbSearchResultItem.java
@@ -42,7 +42,7 @@ public class TmdbSearchResultItem {
     private Double voteAverage;
 
     @JsonProperty("vote_count")
-    private Integer voteCount;
+    private Long voteCount;
 
     private Double popularity;
